Move job handler's pre-checks in a new class
Job handlers should be retriable and may have a CeleryTask reference.
If a job handler is instantiated just for calling the pre-check method,
then it can not be given a CeleryTask reference.

The pre_check method is now a classmethod for any job handler.
The job handler's pre_check method will invoke the pre_check method of
a list of new checker classes.
A checker class needs all the parameters needed by a job handler class
but they don't need the CeleryTask reference.

All the common code between ceckers and handlers has been moved
in mixin classes.
