+++ packit_service/worker/handlers/koji.py
@@ -7,7 +7,7 @@ This file defines classes for job handlers specific for Fedmsg events
 
 import logging
 from datetime import datetime
-from typing import Optional
+from typing import Optional, Tuple, Type
 
 from ogr.abstract import GitProject
 from packit.config import (
@@ -16,16 +16,14 @@ from packit.config import (
 )
 from packit.config.package_config import PackageConfig
 from packit_service.constants import (
-    KOJI_PRODUCTION_BUILDS_ISSUE,
     KojiBuildState,
-    PERMISSIONS_ERROR_WRITE_OR_ADMIN,
 )
 from packit_service.constants import KojiTaskState
 from packit_service.models import AbstractTriggerDbType, KojiBuildTargetModel
 from packit_service.service.urls import (
     get_koji_build_info_url,
 )
-from packit_service.worker.events.enums import GitlabEventAction
+from packit_service.worker.checker.abstract import Checker
 from packit_service.worker.helpers.build.koji_build import KojiBuildJobHelper
 from packit_service.worker.events import (
     CheckRerunCommitEvent,
@@ -36,7 +34,6 @@ from packit_service.worker.events import (
     PullRequestGithubEvent,
     PushGitHubEvent,
     PushGitlabEvent,
-    PushPagureEvent,
     ReleaseEvent,
     AbstractPRCommentEvent,
 )
@@ -51,6 +48,8 @@ from packit_service.worker.handlers.abstract import (
 )
 from packit_service.worker.reporting import BaseCommitStatus
 from packit_service.worker.result import TaskResults
+from packit_service.worker.checker import koji
+from packit_service.worker.handlers.mixin import GetKojiBuildJobHelperMixin
 
 logger = logging.getLogger(__name__)
 
@@ -70,7 +69,7 @@ logger = logging.getLogger(__name__)
 @reacts_to(CheckRerunPullRequestEvent)
 @reacts_to(CheckRerunCommitEvent)
 @reacts_to(CheckRerunReleaseEvent)
-class KojiBuildHandler(JobHandler):
+class KojiBuildHandler(JobHandler, GetKojiBuildJobHelperMixin):
     task_name = TaskName.upstream_koji_build
 
     def __init__(
@@ -89,64 +88,12 @@ class KojiBuildHandler(JobHandler):
         self._koji_build_helper: Optional[KojiBuildJobHelper] = None
         self._project: Optional[GitProject] = None
 
-    @property
-    def koji_build_helper(self) -> KojiBuildJobHelper:
-        if not self._koji_build_helper:
-            self._koji_build_helper = KojiBuildJobHelper(
-                service_config=self.service_config,
-                package_config=self.package_config,
-                project=self.project,
-                metadata=self.data,
-                db_trigger=self.data.db_trigger,
-                job_config=self.job_config,
-                build_targets_override=self.data.build_targets_override,
-                tests_targets_override=self.data.tests_targets_override,
-            )
-        return self._koji_build_helper
+    @staticmethod
+    def get_checkers() -> Tuple[Type[Checker]]:
+        return (koji.Permission,)
 
     def run(self) -> TaskResults:
         return self.koji_build_helper.run_koji_build()
-
-    def pre_check(self) -> bool:
-        if (
-            self.data.event_type == MergeRequestGitlabEvent.__name__
-            and self.data.event_dict["action"] == GitlabEventAction.closed.value
-        ):
-            # Not interested in closed merge requests
-            return False
-
-        if self.data.event_type in (
-            PushGitHubEvent.__name__,
-            PushGitlabEvent.__name__,
-            PushPagureEvent.__name__,
-        ):
-            configured_branch = self.koji_build_helper.job_build_branch
-            if self.data.git_ref != configured_branch:
-                logger.info(
-                    f"Skipping build on '{self.data.git_ref}'. "
-                    f"Push configured only for '{configured_branch}'."
-                )
-                return False
-
-        if self.data.event_type == PullRequestGithubEvent.__name__:
-            user_can_merge_pr = self.project.can_merge_pr(self.data.actor)
-            if not (user_can_merge_pr or self.data.actor in self.service_config.admins):
-                self.koji_build_helper.report_status_to_all(
-                    description=PERMISSIONS_ERROR_WRITE_OR_ADMIN,
-                    state=BaseCommitStatus.neutral,
-                )
-                return False
-
-        if not self.koji_build_helper.is_scratch:
-            msg = "Non-scratch builds not possible from upstream."
-            self.koji_build_helper.report_status_to_all(
-                description=msg,
-                state=BaseCommitStatus.neutral,
-                url=KOJI_PRODUCTION_BUILDS_ISSUE,
-            )
-            return False
-
-        return True
 
 
 @configured_as(job_type=JobType.production_build)
